new: split tests in multiple files
