+++ tests/StateDecorator.test.tsx
@@ -1357,385 +1357,6 @@ describe('StateDecorator', () => {
     });
   });
 
-  describe('Conflicting actions management', () => {
-    it('handles 2 calls to an asynchronous action (keep_all + null in 2nd call)', (done) => {
-      type State = { count: number };
-      type Actions = {
-        getData: (value: string) => Promise<string>;
-      };
-
-      let count = 1;
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        getData: {
-          promise: ([value]) => {
-            if (count === 1 || count === 3) {
-              count = 2;
-              return new Promise((res) => {
-                setTimeout(res, 200, value);
-              });
-            }
-            count++;
-            return null;
-          },
-          reducer: (s): State => ({
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.KEEP_ALL,
-        },
-      };
-
-      const onMount = jestFail(done, (actions: Actions) => {
-        actions.getData('id1').catch((e) => done.fail(e));
-        actions.getData('id1').catch((e) => done.fail(e));
-        actions
-          .getData('id1')
-          .then(() => {
-            done();
-          })
-          .catch((e) => done.fail(e));
-      });
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, () => <div />);
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 2 calls to an asynchronous action (reuse)', (done) => {
-      type State = { count: number };
-      type Actions = {
-        getData: (value: string) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        getData: {
-          promise: ([value]) => new Promise((res) => setTimeout(res, 200, value)),
-          reducer: (s): State => ({
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.REUSE,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        const p1 = actions.getData('id1').catch((e) => done.fail(e));
-        const p2 = actions.getData('id1').catch((e) => done.fail(e));
-        expect(p1 === p2);
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        if (state.count === 1) {
-          done();
-        }
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 2 calls to an asynchronous action (reuse)', (done) => {
-      type State = { count: number; res: string };
-      type Actions = {
-        getData: (value: string) => Promise<string>;
-      };
-
-      let timeout = 100;
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        getData: {
-          promise: ([value]) =>
-            new Promise((res) => {
-              setTimeout(res, timeout, value);
-              timeout = timeout + 200;
-            }),
-          reducer: (s, res): State => ({
-            res,
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.REUSE,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        const p1 = actions.getData('id1').catch((e) => done.fail(e));
-        const p2 = actions.getData('id2').catch((e) => done.fail(e));
-        expect(p1 !== p2);
-        p2.then(() => {
-          done();
-        });
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 3 calls to an asynchronous action (keep last)', (done) => {
-      type State = { str: string; count: number };
-      type Actions = {
-        asynch: (value: string, timeout: number) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        asynch: {
-          promise: ([value, timeout]) => new Promise((res) => setTimeout(res, timeout, value)),
-          reducer: (s, res): State => ({
-            str: res,
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.KEEP_LAST,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        actions.asynch('value 1', 200).catch((e) => done.fail(e));
-        actions.asynch('value 2', 0).catch((e) => done.fail(e));
-        actions.asynch('value 3', 0).catch((e) => done.fail(e));
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          str: 'init',
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        if (state.str === 'value 3') {
-          expect(state.count).toBe(2);
-          done();
-        }
-
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 3 calls to an asynchronous action (keep all)', (done) => {
-      type State = { str: string; count: number };
-      type Actions = {
-        asynch: (value: string, timeout: number) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        asynch: {
-          promise: ([value, timeout]) => new Promise((res) => setTimeout(res, timeout, value)),
-          reducer: (s, res): State => ({
-            str: res,
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.KEEP_ALL,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        actions.asynch('value 1', 200).catch((e) => done.fail(e));
-        actions.asynch('value 2', 100).catch((e) => done.fail(e));
-        actions.asynch('value 3', 0).catch((e) => done.fail(e));
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          str: 'init',
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        if (state.str === 'value 3') {
-          expect(state.count).toBe(3);
-          done();
-        }
-
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 3 calls to an asynchronous action (ignore)', (done) => {
-      type State = { str: string; count: number };
-      type Actions = {
-        asynch: (value: string, timeout: number) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        asynch: {
-          promise: ([value, timeout]) => new Promise((res) => setTimeout(res, timeout, value)),
-          reducer: (s, res): State => ({
-            str: res,
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.IGNORE,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        actions.asynch('value 1', 200).catch((e) => done.fail(e));
-        actions.asynch('value 2', 0).catch((e) => done.fail(e));
-        actions.asynch('value 3', 0).catch((e) => done.fail(e));
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          str: 'init',
-          count: 0,
-        } as State,
-      };
-
-      let saveState: State;
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        saveState = state;
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-
-      return new Promise((res) => {
-        setTimeout(() => {
-          expect(saveState.count).toBe(1);
-          expect(saveState.str).toBe('value 1');
-          res();
-          done();
-        }, 300);
-      });
-    });
-
-    it('handles 3 calls to an asynchronous action (reject)', (done) => {
-      type State = { str: string; count: number };
-      type Actions = {
-        asynch: (value: string, timeout: number) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        asynch: {
-          promise: ([value, timeout]) => new Promise((res) => setTimeout(res, timeout, value)),
-          reducer: (s, res): State => ({
-            str: res,
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.REJECT,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        actions.asynch('value 1', 200).catch((e) => done.fail(e));
-        actions.asynch('value 2', 0).catch((e) => done());
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          str: 'init',
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap) => {
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-
-    it('handles 3 calls to an asynchronous action (parallel)', (done) => {
-      type State = { users: { [k: string]: string }; count: number };
-      type Actions = {
-        asynch: (userId: string, value: string, timeout: number) => Promise<string>;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        asynch: {
-          promise: (userId, value, timeout) => new Promise((res) => setTimeout(res, timeout, value)),
-          reducer: (s, res, [userId, value]): State => ({
-            users: {
-              ...s.users,
-              [userId]: value,
-            },
-            count: s.count + 1,
-          }),
-          conflictPolicy: ConflictPolicy.PARALLEL,
-          getPromiseId: (userId) => userId,
-        },
-      };
-
-      const onMount = (actions: Actions) => {
-        Promise.all([
-          actions.asynch('user1', 'value 1', 100).catch((e) => done.fail(e)),
-          actions.asynch('user2', 'value 2', 300).catch((e) => done.fail(e)),
-          actions.asynch('user3', 'value 3', 200).catch((e) => done.fail(e)),
-        ]).then(() => {
-          setTimeout(() => {
-            actions.asynch('user4', 'value 4', 100).catch((e) => done.fail(e));
-            actions.asynch('user5', 'value 5', 300).catch((e) => done.fail(e));
-          }, 300);
-        });
-      };
-
-      const props = {
-        actions,
-        onMount,
-        initialState: {
-          users: {},
-          count: 0,
-        } as State,
-      };
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap, loadingParallelMap) => {
-        if (state.count > 0 && state.count !== 5) {
-          if (!loadingParallelMap.asynch['user1']) {
-            expect(state.users['user1']).toEqual('value 1');
-          }
-        }
-        if (loadingMap.asyncAction) {
-          expect(Object.keys(loadingParallelMap.asyncAction).length).toBeGreaterThan(0);
-        }
-        if (state.count === 5) {
-          done();
-        }
-
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-    });
-  });
-
   it('exposes the actions in the Asynch action', (done) => {
     const notifySuccess = jest.fn();
 
@@ -1824,205 +1445,6 @@ describe('StateDecorator', () => {
 
     shallow(<StateDecorator {...props}>{() => <div />}</StateDecorator>);
   });
-
-  describe('On props change', () => {
-    it('when no props of interest change, callbacks are not called 1', () => {
-      type Item = {
-        id: string;
-        value: string;
-      };
-
-      type State = {
-        item: Item;
-      };
-
-      const initialState: State = {
-        item: {
-          id: 'id1',
-          value: 'value1',
-        },
-      };
-
-      type Actions = {
-        get: () => any;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        get: (s) => s,
-      };
-
-      const props: StateDecoratorProps<State, Actions, ItemMap> = {
-        actions,
-        initialState,
-        getPropsRefValues: (p) => [],
-        onPropsChange: jest.fn(),
-        onPropsChangeReducer: jest.fn(),
-        props: {
-          item: {
-            id: 'id1',
-            value: 'value1',
-          },
-        },
-      };
-
-      const wrapper = shallow(<StateDecorator {...props}>{() => <div />}</StateDecorator>);
-
-      wrapper.setProps({
-        item: {
-          id: 'id2',
-          value: 'value2',
-        },
-      });
-
-      expect(props.onPropsChange).not.toHaveBeenCalled();
-      expect(props.onPropsChangeReducer).not.toHaveBeenCalled();
-    });
-
-    it('when no props of interest change, callbacks are not called 2', () => {
-      type Item = {
-        id: string;
-        value: string;
-      };
-
-      type State = {
-        item: Item;
-        item2: Item;
-      };
-
-      const initialState: State = {
-        item: {
-          id: 'id1',
-          value: 'value1',
-        },
-        item2: {
-          id: 'id1',
-          value: 'value1',
-        },
-      };
-
-      type Actions = {
-        get: () => any;
-      };
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        get: (s) => s,
-      };
-
-      const props: StateDecoratorProps<State, Actions, ItemMap> = {
-        actions,
-        initialState,
-        getPropsRefValues: (p) => [p.item.id],
-        onPropsChange: jest.fn(),
-        onPropsChangeReducer: jest.fn(),
-        props: {
-          item: {
-            id: 'id1',
-            value: 'value1',
-          },
-        },
-      };
-
-      const wrapper = shallow(<StateDecorator {...props}>{() => <div />}</StateDecorator>);
-
-      wrapper.setProps({
-        item2: {
-          id: 'id2',
-          value: 'value2',
-        },
-      });
-
-      expect(props.onPropsChange).not.toHaveBeenCalled();
-      expect(props.onPropsChangeReducer).not.toHaveBeenCalled();
-    });
-
-    it('when one prop of interest change, callbacks are called', (done) => {
-      type Item = {
-        id: string;
-        value: string;
-      };
-
-      type State = {
-        item: Item;
-        item2: Item;
-      };
-
-      const initialState: State = {
-        item: {
-          id: 'id1',
-          value: 'value1',
-        },
-        item2: {
-          id: 'id1',
-          value: 'value1',
-        },
-      };
-
-      type Actions = {
-        get: () => any;
-        action2: () => any;
-      };
-
-      const action2Mock = jest.fn();
-
-      const actions: StateDecoratorActions<State, Actions, any> = {
-        get: (s) => s,
-        action2: (s) => {
-          action2Mock();
-          return s;
-        },
-      };
-
-      const props: StateDecoratorProps<State, Actions, ItemMap> = {
-        actions,
-        initialState,
-        getPropsRefValues: (p) => [p && p.item.id, p && p.item2.id],
-        onPropsChange: jest.fn((s, p, a, indices) => {
-          expect(s.item.id).toEqual('id2');
-          expect(p.item.id).toEqual('id2');
-          expect(a.get).toBeInstanceOf(Function);
-          expect(indices).toEqual([0]);
-          a.action2();
-          done();
-        }),
-        onPropsChangeReducer: jest.fn((s, p) => ({ ...s, item: p.item })),
-        props: {
-          item: {
-            id: 'id1',
-            value: 'value1',
-          },
-          item2: {
-            id: 'id1',
-            value: 'value1',
-          },
-        },
-      };
-
-      let stateHasChanged = false;
-
-      const renderFunction = jestFail(done, (state: State, actions, loading, loadingMap, loadingParallelMap) => {
-        if (state.item.id === 'id2') {
-          stateHasChanged = true;
-        }
-
-        return <div />;
-      });
-
-      const wrapper = shallow(<StateDecorator {...props}>{renderFunction}</StateDecorator>);
-
-      wrapper.setProps({
-        props: {
-          ...props.props,
-          item: {
-            id: 'id2',
-            value: 'value2',
-          },
-        },
-      });
-
-      expect(props.onPropsChangeReducer).toHaveBeenCalled();
-      expect(stateHasChanged).toBeTruthy();
-    });
-  });
 });
 
 describe('Type guards', () => {
@@ -2050,222 +1472,5 @@ describe('Type guards', () => {
   it('test synchronous actions correctly', () => {
     expect(isSyncAction(actions.syncAction)).toBeTruthy();
     expect(isAsyncAction(actions.syncAction)).toBeFalsy();
-  });
-});
-
-describe('retryDecorator', () => {
-  it('3 retries, immediate success', () => {
-    const arg1 = 'arg1';
-    const arg2 = 'arg2';
-
-    const f: PromiseProvider<any, any, any, any> = ([arg1, arg2]) => {
-      expect(arg1).toBe(arg1);
-      expect(arg2).toBe(arg2);
-      return Promise.resolve('OK');
-    };
-
-    return retryDecorator(f, 3)([arg1, arg2], null, null, null);
-  });
-
-  it('3 retries, 1 error, 1 success', () => {
-    const arg1 = 'arg1';
-    const arg2 = 'arg2';
-
-    let count = 1;
-    const f = (arg1, arg2) => {
-      expect(arg1).toBe(arg1);
-      expect(arg2).toBe(arg2);
-      if (count++ === 1) {
-        return Promise.reject(new TypeError());
-      }
-      return Promise.resolve('OK');
-    };
-
-    return retryDecorator(f, 3, 0)([arg1, arg2], null, null, null).then((res) => expect(res).toBe('OK'));
-  });
-
-  it('3 retries, 3 failures', (done) => {
-    const arg1 = 'arg1';
-    const arg2 = 'arg2';
-
-    const f = (arg1, arg2) => {
-      expect(arg1).toBe(arg1);
-      expect(arg2).toBe(arg2);
-      return Promise.reject(new TypeError());
-    };
-
-    return retryDecorator(f, 3, 0)([arg1, arg2], null, null, null)
-      .then(() => done.fail())
-      .catch((e) => {
-        expect(e).toBeInstanceOf(TypeError);
-        done();
-      });
-  });
-
-  it('3 retries, 1 different failure', (done) => {
-    const arg1 = 'arg1';
-    const arg2 = 'arg2';
-
-    const f = (arg1, arg2) => {
-      expect(arg1).toBe(arg1);
-      expect(arg2).toBe(arg2);
-      throw new Error();
-    };
-
-    return retryDecorator(f, 3)([arg1, arg2], null, null, null)
-      .then(() => done.fail())
-      .catch((e) => {
-        expect(e).not.toBeInstanceOf(TypeError);
-        done();
-      });
-  });
-
-  it('3 retries, 1 retry error', (done) => {
-    const arg1 = 'arg1';
-    const arg2 = 'arg2';
-
-    const isRetryError = (e: Error) => {
-      return e.message === 'OK';
-    };
-
-    let count = 1;
-    const f = (arg1, arg2) => {
-      expect(arg1).toBe(arg1);
-      expect(arg2).toBe(arg2);
-      if (count++ === 1) {
-        return Promise.reject(new Error('OK'));
-      }
-      return Promise.resolve('OK');
-    };
-
-    return retryDecorator(f, 3, 0, isRetryError)([arg1, arg2], null, null, null)
-      .then((res) => {
-        expect(res).toEqual('OK');
-        done();
-      })
-      .catch(done.fail);
-  });
-
-  describe('Testing utilities', () => {
-    type State = {
-      value: number;
-    };
-    type Actions = {
-      increment: (v: number) => void;
-      incrAsync: (v: number) => Promise<number>;
-      incrComplexSync: (v: number) => Promise<number>;
-      incrAsyncGet: (v: number) => Promise<number>;
-    };
-    type Props = {};
-
-    const actions: StateDecoratorActions<State, Actions, Props> = {
-      increment: (s, [incr], props) => {
-        return {
-          value: s.value + incr,
-        };
-      },
-      incrComplexSync: {
-        action: (s, [incr], props) => ({
-          value: s.value + incr,
-        }),
-      },
-      incrAsync: {
-        promise: ([incr]) => Promise.resolve(incr),
-        reducer: (s, incr) => ({ value: s.value + incr }),
-      },
-      incrAsyncGet: {
-        promiseGet: ([incr]) => Promise.resolve(incr),
-        reducer: (s, incr) => ({ value: s.value + incr }),
-      },
-    };
-
-    it('testSyncAction (correct type)', (done) => {
-      const p = testSyncAction(actions.increment, (action) => {
-        expect(action).toBeDefined();
-        done();
-      });
-
-      p && p.catch((e) => done.fail(e));
-    });
-
-    it('testSyncAction (incorrect type)', (done) => {
-      const testFunc = jest.fn();
-
-      testSyncAction(actions.incrAsync, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
-
-    it('testSyncAction (incorrect type 2)', (done) => {
-      const testFunc = jest.fn();
-
-      testSyncAction(actions.incrComplexSync, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
-
-    it('testAsyncAction (correct type)', (done) => {
-      const p = testAsyncAction(actions.incrAsync, (action) => {
-        expect(action).toBeDefined();
-        done();
-      });
-      p && p.catch((e) => done.fail(e));
-    });
-
-    it('testAsyncAction (promiseGet converted in promise)', (done) => {
-      const p = testAsyncAction(actions.incrAsyncGet, (action) => {
-        expect(action).toBeDefined();
-        expect(action.promise).toBeDefined();
-        done();
-      });
-      p && p.catch((e) => done.fail(e));
-    });
-
-    it('testAsyncAction (incorrect type)', (done) => {
-      const testFunc = jest.fn();
-
-      testAsyncAction(actions.increment, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
-
-    it('testAsyncAction (incorrect type2)', (done) => {
-      const testFunc = jest.fn();
-
-      testAsyncAction(actions.incrComplexSync, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
-
-    it('testSyncComplexAction (promiseGet converted in promise)', (done) => {
-      const p = testAdvancedSyncAction(actions.incrComplexSync, (action) => {
-        expect(action).toBeDefined();
-        expect(action.action).toBeDefined();
-        done();
-      });
-      p && p.catch((e) => done.fail(e));
-    });
-
-    it('testSyncComplexAction (incorrect type)', (done) => {
-      const testFunc = jest.fn();
-
-      testAdvancedSyncAction(actions.increment, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
-
-    it('testSyncComplexAction (incorrect type2)', (done) => {
-      const testFunc = jest.fn();
-
-      testAdvancedSyncAction(actions.incrAsync, testFunc).catch((e) => {
-        expect(testFunc).not.toHaveBeenCalled();
-        done();
-      });
-    });
   });
 });
